chore(build): Use webpack to build
Switch to use babel for development, and webpack to create a
distrobution file.

Now installable through npm and bower. Bower requires the dist
files to be part of the repository...

BREAKING CHANGE:

Runtime file to be included is moved to `dist/bim-select.js`.
And the CSS file is part of the javascript now, so no need to
include it separately.

Fixes #5.
Fixes #2.