feat(selected item template): introduces a template for the selected item.
This allows you to render html where you would normally only be able to render text in the input field.

BREAKING CHANGE: the template variable `match` in the bim-select-item
template has changed to `item`. This was done so that the bim-select-item
template can be reused for the bim-select-selected-item template where
the variable name `match` does not really make sense since the selected item
is no longer a match but rather is the selected item after the match item was
selected.